requirements.txt
crewai
crewai-tools
fastapi==0.110.3
uvicorn==0.30.1
python-dotenv==1.0.1
openai>=1.68.2,<2
pypdf
langchain
chromadb